fix loc.end info for multi-line JSXAttributeValue
on the old code, the updated test would fail since `loc.end` would contain
`{ line: 1, column: 62 }`

also pull new-line handling code out of readTmplToken for easy reuse. note
that we don't do any normalization of `<CR><LF>` in a JSXAttributeValue,
so as to be consistent with esprima-fb.

it's worth pointing out that per the spec, a standalone `<CR>` inside a
TemplateLiteral should also be normalized to a `<LF>`, but i think that
warrants a separate discussion/PR. this commit doesn't change any
TemplateLiteral behavior.